add ability to delete EnemyAgent
We will want to be able to dynamically create and delete enemy agnets
throughout the game. The delete part is hard with shared pointers. Essneitally
the GameController is the owner of the EnemyAgents and is responsible
for createing and destroying them based on what happens in the game, as it facilitates
everything. Hence chagend to raw pointer approach. Additionally only store
a reference to the contianer holding the enemy agent ptrs at the game GameModel
level as it needs to access them but does not own them.