update game controller to take shared pointer
Using a shared pointer to a container of static objects is advantageous
in that:

If the container is updated, all entities that have a reference/own the
shared pointer see that container with the update.

We can delete static objects (EnemyAgents) off the container itself as
although the container is a dynamic vector, the values inside the
container are static objects.